feat: refactor StepBarView and introduce StepBarOption interface for improved structure and initialization
