+++ src/stepBar/index.ts
@@ -1,1 +1,2 @@
 export * from "./StepBarView.js";
+export * from "./StepBarOption.js";
